requirements.txt
addict
future
lmdb
numpy
opencv-python
pandas
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
timm
torch>=1.8.1
torchvision>=0.9
tqdm
yapf
